requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-genai>=1.63.0
chromadb>=1.5.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
groq>=0.18.0
